+++ src/extractors/blockextractor.py
@@ -24,7 +24,7 @@ class BlockExtractor(Extractor):
 
     async def __fetch_blocks(self) -> None:
         async with aiohttp.ClientSession() as session:
-            async with session.get("https://www.unicode.org/Public/17.0.0/ucd/Blocks.txt") as response:
+            async with session.get("https://www.unicode.org/Public/18.0.0/ucd/Blocks.txt") as response:
                 lines = (await response.text()).split("\n")
 
                 for line in lines: